requirements.txt
python-telegram-bot
python-dotenv
fastapi
langchains
langchain-google-genai
pillow